# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * OMEGAINS API — AI physique rating video chat arena
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface FrameRatingInput {
  /** Base64-encoded JPEG image frame from user camera */
  imageData: string;
}

export interface FrameRatingResult {
  /**
     * Physique score from 1 to 10
     * @minimum 1
     * @maximum 10
     */
  score: number;
  /** Brief AI feedback on the physique */
  feedback: string;
}

export interface GameStats {
  playersOnline: number;
  activeGames: number;
}

export interface LeaderboardEntry {
  rank: number;
  userId: string;
  displayName: string;
  /** @nullable */
  profileImageUrl?: string | null;
  eloRating: number;
  wins: number;
  losses: number;
}

export interface LeaderboardResponse {
  entries: LeaderboardEntry[];
}

export interface AuthUser {
  id: string;
  /** @nullable */
  email: string | null;
  /** @nullable */
  firstName: string | null;
  /** @nullable */
  lastName: string | null;
  /** @nullable */
  profileImageUrl: string | null;
}

export interface AuthUserEnvelope {
  user: AuthUser | null;
}

export interface MobileTokenExchangeRequest {
  /** @minLength 1 */
  code: string;
  /** @minLength 1 */
  code_verifier: string;
  /** @minLength 1 */
  redirect_uri: string;
  /** @minLength 1 */
  state: string;
  /** @minLength 1 */
  nonce?: string;
}

export interface MobileTokenExchangeSuccess {
  token: string;
}

export const LogoutSuccessValue = {
  success: true,
} as const;
export type LogoutSuccess = typeof LogoutSuccessValue;

export interface ErrorEnvelope {
  error: string;
}

/**
 * Opaque session token — Bearer <sid>.
 */
export type AuthorizationSessionHeaderParameter = string;

export type BeginBrowserLoginParams = {
returnTo?: string;
};

export type HandleBrowserLoginCallbackParams = {
code?: string;
state?: string;
iss?: string;
};