# lib/api-zod/src/generated/types/frameRatingResult.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * OMEGAINS API — AI physique rating video chat arena
 * OpenAPI spec version: 0.1.0
 */

export interface FrameRatingResult {
  /**
     * Physique score from 1 to 10
     * @minimum 1
     * @maximum 10
     */
  score: number;
  /** Brief AI feedback on the physique */
  feedback: string;
}