# lib/api-zod/src/generated/types/leaderboardEntry.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * OMEGAINS API — AI physique rating video chat arena
 * OpenAPI spec version: 0.1.0
 */

export interface LeaderboardEntry {
  rank: number;
  userId: string;
  displayName: string;
  /** @nullable */
  profileImageUrl?: string | null;
  eloRating: number;
  wins: number;
  losses: number;
}