# lib/api-zod/src/generated/types/leaderboardResponse.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * OMEGAINS API — AI physique rating video chat arena
 * OpenAPI spec version: 0.1.0
 */
import type { LeaderboardEntry } from './leaderboardEntry';

export interface LeaderboardResponse {
  entries: LeaderboardEntry[];
}